`TransactionsManager`: always use `SK1` implementation
Fixes [CSDK-513].

The SDK still uses `SK1` when it comes to fetching receipts to post to the backend.
The `SK1` implementation checks whether there are transactions inside of the receipt.

However, this could lead to issues if `SK2` has transactions but SK1 doesn't. In those cases, we would end up posting an empty receipt.
This fixes that.
